requirements.txt
fastapi
uvicorn
requests
pydantic[email]
python-multipart
python-dotenv

